requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.24.2
azure-functions==1.11.2
azure-functions-durable==1.1.5
azure-storage-blob==12.13.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
frozenlist==1.3.0
furl==2.1.3
idna==3.3
isodate==0.6.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.0
orderedmultidict==1.0.1
pycparser==2.21
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
yarl==1.7.2
